merge-tree.c: support --merge-base in conjunction with --stdin
The previous change add "--merge-base" option in order to allow users to
specify merge-base for the merge. But it doesn't compatible well with
--stdin, because multiple batched merges can only have the same specified
base.

This patch allows users to pass --merge-base option into the input line,
such as:

    printf "--merge-base=<b3> <b1> <b2>" | git merge-tree --stdin

This does a merge of b1 and b2, and uses b3 as the merge-base.

Signed-off-by: Kyle Zhao <kylezhao@tencent.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>